+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+### New features
+
+* [#42](https://github.com/rubocop-hq/rubocop-rails/issues/42): Add new `Rails/CredentialsDirectAccess` cop. ([@tejasbubane][])
+
 ## 2.0.0 (2019-05-22)
 
 ### New features
@@ -11,3 +15,4 @@
 
 [@koic]: https://github.com/koic
 [@andyw8]: https://github.com/andyw8
+[@tejasbubane]: https://github.com/tejasbubane